+++ src/app/onboarding/components/OnboardingFunnel.tsx
@@ -1,35 +1,20 @@
 'use client';
 
-import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
-import { KeyboardEvent, useEffect, useState } from 'react';
-import { toast } from 'react-toastify';
-import { OnboardingFormValues } from '@/components/onboarding-contents/onboarding.type';
 import { useRouter } from 'next/navigation';
-import { FORM_DEFAULT_VALUES } from '@/components/onboarding-contents/formValidation.const';
-import PersonalInfoStep from './steps/PersonalInfoStep';
-import { usePutUser } from '@/services/user.service';
-import { AgeRange, Gender, UpdateMeRequest } from '@/types/user.type';
 import { setOnboardingStatusComplete } from '@/utils/handleToken.util';
-import useFunnel from '@/hooks/useFunnel';
 import PhoneNumberStep from './steps/PhoneNumberStep';
-
-const ONBOARDING_STEPS = ['phoneNumber', 'personalInfo'] as const;
+import { useEffect } from 'react';
+import { FormProvider, useForm } from 'react-hook-form';
+import { FORM_DEFAULT_VALUES } from '@/components/onboarding-contents/formValidation.const';
+import { OnboardingFormValues } from '@/components/onboarding-contents/onboarding.type';
 
 interface Props {
   isOnboardingComplete: boolean;
-  initialPhoneNumber: string;
-  initialName: string;
-  initialGender: Exclude<Gender, 'NONE'> | null;
-  initialAgeRange: Exclude<AgeRange, '연령대 미지정'> | null;
 }
 
-const OnboardingFunnel = ({
-  isOnboardingComplete,
-  initialPhoneNumber,
-  initialName,
-  initialGender,
-  initialAgeRange,
-}: Props) => {
+const OnboardingFunnel = ({ isOnboardingComplete }: Props) => {
+  const router = useRouter();
+
   useEffect(() => {
     if (isOnboardingComplete) {
       setOnboardingStatusComplete();
@@ -40,90 +25,14 @@ const OnboardingFunnel = ({
   const methods = useForm<OnboardingFormValues>({
     defaultValues: {
       ...FORM_DEFAULT_VALUES,
-      phoneNumber: initialPhoneNumber,
-      name: initialName,
-      gender:
-        initialGender === 'MALE'
-          ? '남성'
-          : initialGender === 'FEMALE'
-            ? '여성'
-            : undefined,
-      age: initialAgeRange ?? undefined,
-    },
-    mode: 'onBlur',
-  });
-
-  // const initialStep = initialPhoneNumber ? 'personalInfo' : 'phoneNumber';
-  const initialStep = 'phoneNumber';
-
-  const { Funnel, Step, handleNextStep } = useFunnel(
-    ONBOARDING_STEPS,
-    initialStep,
-  );
-
-  const router = useRouter();
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const { mutate: putUser, isSuccess } = usePutUser({
-    options: {
-      skipCheckOnboarding: true,
-    },
-    onSuccess: async () => {
-      setOnboardingStatusComplete();
-      router.replace('/');
-    },
-    onError: (e) => {
-      console.error(e);
-      toast.error('회원가입에 실패했어요.');
-      setIsSubmitting(false);
     },
   });
-
-  const submitForm: SubmitHandler<OnboardingFormValues> = async (formData) => {
-    setIsSubmitting(true);
-    const body: UpdateMeRequest = {
-      phoneNumber: formData.phoneNumber,
-      name: formData.name,
-      ageRange: formData.age,
-      gender:
-        formData.gender === '남성' ? ('MALE' as const) : ('FEMALE' as const),
-    };
-    putUser(body);
-  };
-  const triggerSubmitForm = () => {
-    const onboardingForm = document.getElementById(
-      'onboarding-form',
-    ) as HTMLFormElement;
-    onboardingForm?.requestSubmit();
-  };
-  const handleEnter = (e: KeyboardEvent<HTMLFormElement>) => {
-    if (e.key === 'Enter') {
-      e.preventDefault();
-    }
-  };
-
-  const disabled = isSubmitting || isSuccess;
 
   return (
     <FormProvider {...methods}>
-      <form
-        onSubmit={methods.handleSubmit(submitForm)}
-        onKeyDown={handleEnter}
-        noValidate
-        className="relative flex w-full grow flex-col"
-        id="onboarding-form"
-      >
-        <Funnel>
-          <Step name="phoneNumber">
-            <PhoneNumberStep handleNextStep={handleNextStep} />
-          </Step>
-          <Step name="personalInfo">
-            <PersonalInfoStep
-              triggerSubmitForm={triggerSubmitForm}
-              disabled={disabled}
-            />
-          </Step>
-        </Funnel>
-      </form>
+      <main className="relative flex w-full grow flex-col">
+        <PhoneNumberStep />
+      </main>
     </FormProvider>
   );
 };
