+++ src/types/user.type.ts
@@ -68,7 +68,6 @@ export const UpdateMeRequestSchema = z
     profileImage: z.string().nullable(),
     name: z.string(),
     nickname: z.string(),
-    phoneNumber: z.string(),
     gender: GenderEnum.exclude(['NONE']),
     ageRange: AgeRangeEnum.exclude(['연령대 미지정']),
     regionId: z.string(),
